@@ -4,11 +4,21 @@ import { BrowserModule } from "@angular/platform-browser";
 import { AppComponent } from "./app.component";
 import { CurrentWeatherComponent } from "./current-weather/current-weather.component";
 import { WeatherService } from "./weather/weather.service";
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { MaterialModule } from './material.module';
+import 'hammerjs';
+import { FlexLayoutModule } from '@angular/flex-layout';
 
 @NgModule({
   declarations: [AppComponent, CurrentWeatherComponent],
-  imports: [BrowserModule, HttpClientModule],
+  imports: [
+    BrowserModule,
+    HttpClientModule,
+    BrowserAnimationsModule,
+    MaterialModule,
+    FlexLayoutModule
+  ],
   providers: [WeatherService],
   bootstrap: [AppComponent]
 })
-export class AppModule {}
+export class AppModule { }
