@@ -3,10 +3,13 @@ import { AppComponent } from "./app.component";
 import { CurrentWeatherComponent } from "./current-weather/current-weather.component";
 import { WeatherService } from "./weather/weather.service";
 import { WeatherServiceFake } from "./weather/weather.service.fake";
+import { MaterialModule } from "./material.module";
+import { NoopAnimationsModule } from "@angular/platform-browser/animations";
 describe("AppComponent", () => {
   beforeEach(async(() => {
     TestBed.configureTestingModule({
       declarations: [AppComponent, CurrentWeatherComponent],
+      imports: [MaterialModule, NoopAnimationsModule],
       providers: [{ provide: WeatherService, useClass: WeatherServiceFake }]
     }).compileComponents();
   }));
@@ -19,6 +22,6 @@ describe("AppComponent", () => {
     const fixture = TestBed.createComponent(AppComponent);
     fixture.detectChanges();
     const compiled = fixture.debugElement.nativeElement;
-    expect(compiled.querySelector("h1").textContent).toContain("Local Weather");
+    expect(compiled.querySelector("span").textContent).toContain("Local Weather");
   }));
 });
