@@ -10,11 +10,18 @@ import { WeatherService } from "../weather/weather.service";
 export class CurrentWeatherComponent implements OnInit {
   current: ICurrentWeather;
 
-  constructor(private weatherService: WeatherService) {}
+  constructor(private weatherService: WeatherService) { }
 
   ngOnInit() {
     this.weatherService.getCurrentWeather("Vining", "US").subscribe(data => {
       this.current = data;
     });
+  }
+
+  getOrdinal(date: number) {
+    const n = new Date(date).getDate();
+    return n > 0
+      ? ['th', 'st', 'nd', 'rd'][(n > 3 && n < 21) || n % 10 > 3 ? 0 : n % 10]
+      : '';
   }
 }
