@@ -1,3 +1,5 @@
 export const environment = {
-  production: true
+  production: true,
+  appId: "ff181422e790715d366ac57fa8b6c162",
+  baseUrl: "https://"
 };
